[ISSUE #8212]♻️Extract RocksDB foundation into its owner crate
