+++ docs/plans/architecture-refactor-migration/README.md
@@ -49,7 +49,7 @@
 
 M01 入口有 22 个根 workspace package；M03 加入 `rocketmq-model` 和 `rocketmq-security-api`，M04 加入
 `rocketmq-protocol`，M05 加入 `rocketmq-transport`，M06 capability spike 加入 `rocketmq-store-api`，
-M06-03a leaf foundation 加入 `rocketmq-store-local`，当前为 28 个。
+M06-03a leaf foundation 加入 `rocketmq-store-local`，PR-M06-09 加入 `rocketmq-store-rocksdb`，当前为 29 个。
 以下 10 个新 crate 按计划逐步加入，最终精确达到 32 个：
 
 | 新 crate | 首次创建里程碑 | 最终职责 |
