+++ rocketmq-store-rocksdb/Cargo.toml
@@ -1,0 +1,40 @@
+# Copyright 2023 The RocketMQ Rust Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+[package]
+name = "rocketmq-store-rocksdb"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license.workspace = true
+keywords.workspace = true
+description = "RocksDB storage foundation for Apache RocketMQ in Rust."
+rust-version.workspace = true
+
+[features]
+default = []
+
+[dependencies]
+bytes.workspace = true
+rocksdb.workspace = true
+rocketmq-error.workspace = true
+rocketmq-observability.workspace = true
+rocketmq-runtime.workspace = true
+tokio.workspace = true
+tracing.workspace = true
+
+[dev-dependencies]
+tempfile = "3.27.0"
