+++ rocketmq-store-rocksdb/src/codec.rs
@@ -12,6 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub use rocketmq_store_rocksdb::store::KeyValueStore;
-pub use rocketmq_store_rocksdb::store::RocksDbStore;
-pub use rocketmq_store_rocksdb::store::RocksDbStoreState;
+use rocketmq_error::RocketMQError;
+
+pub trait RocksDbCodec {
+    type Item;
+
+    fn encode(value: &Self::Item, dst: &mut Vec<u8>) -> Result<(), RocketMQError>;
+
+    fn decode(src: &[u8]) -> Result<Self::Item, RocketMQError>;
+}
