+++ rocketmq-store-rocksdb/src/column_family.rs
@@ -12,6 +12,25 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub use rocketmq_store_rocksdb::store::KeyValueStore;
-pub use rocketmq_store_rocksdb::store::RocksDbStore;
-pub use rocketmq_store_rocksdb::store::RocksDbStoreState;
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub enum RocksDbColumnFamily {
+    Default,
+    ConsumeQueueOffset,
+    Timer,
+    Transaction,
+    PopState,
+    Config(String),
+}
+
+impl RocksDbColumnFamily {
+    pub fn name(&self) -> &str {
+        match self {
+            Self::Default => "default",
+            Self::ConsumeQueueOffset => "offset",
+            Self::Timer => "timer",
+            Self::Transaction => "trans",
+            Self::PopState => "popState",
+            Self::Config(name) => name.as_str(),
+        }
+    }
+}
